+++ elements/jsonform/test/cases/index.js
@@ -15,6 +15,8 @@ export { default as loadCustomEditorInterfaceTest } from "./load-custom-editor-i
 export { default as loadSubmitButtonTest } from "./submit-button";
 export { default as loadShowOptInPropertiesTest } from "./load-show-opt-in-properties";
 export { default as loadButtonsEditorTest } from "./load-buttons-editor";
+export { default as loadBinaryCheckboxTest } from "./load-binary-checkbox";
 export { default as loadSpatialValuesTest } from "./load-spatial-values";
 export { default as validationBehaviorTest } from "./validation-behavior";
 export { default as defaultsTest } from "./defaults";
+export { default as loadGridTest, loadGridStrictTest } from "./load-grid";